+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+* Added support for inline clocks & counters (only in edit mode right now).
+
 ## [0.2.4] - 2024-03-06
 
 ### Modified